PPS-63: Refactor Icon
