+++ src/components/atoms/Icon/Icon.stories.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { withKnobs } from '@storybook/addon-knobs';
 import forIn from 'lodash/forIn';
 
-import Icon from '.';
+import Icon from './Icon';
 
 export default {
     title: 'Atoms/Icon',
