+++ src/components/styled-elements/components/index.js
@@ -1,12 +1,14 @@
-import StyledCard from './StyledCard';
 import StyledButton from './StyledButton';
+import StyledCard from './StyledCard';
+import StyledIcon from './StyledIcon';
 import StyledOverlay from './StyledOverlay';
 import StyledPanel from './StyledPanel';
 import StyledTextBox from './StyledTextBox';
 
 export {
-    StyledCard,
     StyledButton,
+    StyledCard,
+    StyledIcon,
     StyledOverlay,
     StyledPanel,
     StyledTextBox,
