+++ src/components/styled-elements/index.js
@@ -1,6 +1,7 @@
 import {
-    StyledCard,
     StyledButton,
+    StyledCard,
+    StyledIcon,
     StyledOverlay,
     StyledPanel,
     StyledTextBox,
@@ -19,8 +20,9 @@ export {
     FlexColumn,
     PaddedWrapper,
     PullRight,
-    StyledCard,
     StyledButton,
+    StyledCard,
+    StyledIcon,
     StyledOverlay,
     StyledPanel,
     StyledTextBox,
